pythonBackend/pyproject.toml
[tool.poetry]
name = "extractor"
version = "0.1.0"
description = "This tool extracts data from long-form media into a comprehensive and insightful summary"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
ansicon = "1.89.0"
appdirs = "1.4.4"
astor = "0.8.1"
async-timeout = "4.0.3"
attrs = "23.1.0"
blessed = "1.20.0"
certifi = "2023.7.22"
charset-normalizer = "3.2.0"
colorama = "0.4.6"
frozenlist = "1.4.0"
git-python = "1.0.3"
gitdb = "4.0.10"
gitpython = "3.1.35"
idna = "3.4"
inquirer = "3.1.3"
jinxed = "1.2.0"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
multidict = "6.0.4"
openai = "0.27.10"
pygments = "2.16.1"
pyreadline3 = "3.4.1"
python-dotenv = "1.0.0"
python-editor = "1.0.4"
readchar = "4.0.5"
regex = "2023.8.8"
requests = "2.31.0"
rich = "13.5.2"
six = "1.16.0"
smmap = "5.0.0"
tiktoken = "0.4.0"
tokentrim = "0.1.10"
tqdm = "4.66.1"
urllib3 = "2.0.4"
wcwidth = "0.2.6"
wget = "3.2"
yarl = "1.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
